document reseed and add a couple of specific tests

+++ fortran_BASGRA_NZ/set_params.f95
@@ -150,7 +150,7 @@ else
 end if
 
 ! reseed parameters
-reseed_harv_delay  = pa(116)  ! number of days to delay harvest after reseed, must be >=1
+reseed_harv_delay  =  INT(pa(116))  ! number of days to delay harvest after reseed, must be >=1
 reseed_LAI         = pa(117)  ! >=0 the leaf area index to set after reseeding, if < 0 then simply use the current LAI
 reseed_TILG2       = pa(118)  ! Non-elongating generative tiller density after reseed if >=0 otherwise use current state of variable
 reseed_TILG1       = pa(119)  ! Elongating generative tiller density after reseed if >=0 otherwise use current state of variable
